refactor: extract catalog module into common root
