+++ lib/common/catalog/repository/catalog_repository.dart
@@ -1,0 +1,5 @@
+import 'package:state_management/common/catalog/model/catalog.dart';
+
+abstract class CatalogRepository {
+  Future<List<Catalog>>? fetch();
+}
