--- lib/common/data/catalog/repository/catalog_repository.dart
@@ -1,5 +1,0 @@
-import 'package:state_management/common/data/catalog/model/catalog.dart';
-
-abstract class CatalogRepository {
-  Future<List<Catalog>>? fetch();
-}
